requirements.txt
lxml==2.3
django-jsonfield==1.0.1
